add util package class to controller package and delete util package, rename test/util package
