+++ src/main/java/controller/CarNamesSpliter.java
@@ -1,4 +1,4 @@
-package util;
+package controller;
 
 import javax.naming.InvalidNameException;
