+++ src/main/java/controller/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package util;
+package controller;
 
 import java.util.ArrayList;
 import java.util.List;
