+++ src/test/java/controller/splitCarNamesTest.java
@@ -1,4 +1,4 @@
-package util;
+package controller;
 
 import org.junit.Test;
